dirty fixes to make it compile.. all files in this commit need fix - mame.c, usrintrf.c: needs update. currently v106 based - datafile.c, m32util.c, treeview.c, properties.c: aspect-related stuffs need fix
git-svn-id: https://mameplus.svn.sourceforge.net/svnroot/mameplus@752 e0215e0a-1e64-f442-9f1d-d2f61a7f4648

+++ trunk/mamep/src/datafile.c
@@ -2359,7 +2359,7 @@ int load_driver_statistics (char *buffer, int bufsize)
 	char name[100];
 	char year[4];
 	int i, n, x, y;
-	int ax, ay;
+	float ax, ay;
 	int all = 0, cl = 0, vec = 0, vecc = 0, neo = 0, neoc = 0;
 	int pch = 0, pchc = 0, deco = 0, decoc = 0, cvs = 0, cvsc = 0, noyear = 0, nobutt = 0, noinput = 0;
 	int vertical = 0, verticalc = 0, horizontal = 0, horizontalc = 0;
